+++ src/dis7/ActionResponsePdu.cpp
@@ -38,23 +38,23 @@ void ActionResponsePdu::setReceivingID(const EntityID& pX) {
   _receivingID = pX;
 }
 
-unsigned int ActionResponsePdu::getRequestID() const { return _requestID; }
+uint32_t ActionResponsePdu::getRequestID() const { return _requestID; }
 
-void ActionResponsePdu::setRequestID(unsigned int pX) { _requestID = pX; }
+void ActionResponsePdu::setRequestID(uint32_t pX) { _requestID = pX; }
 
-unsigned int ActionResponsePdu::getRequestStatus() const {
+uint32_t ActionResponsePdu::getRequestStatus() const {
   return _requestStatus;
 }
 
-void ActionResponsePdu::setRequestStatus(unsigned int pX) {
+void ActionResponsePdu::setRequestStatus(uint32_t pX) {
   _requestStatus = pX;
 }
 
-unsigned int ActionResponsePdu::getNumberOfFixedDatumRecords() const {
+uint32_t ActionResponsePdu::getNumberOfFixedDatumRecords() const {
   return _fixedDatums.size();
 }
 
-unsigned int ActionResponsePdu::getNumberOfVariableDatumRecords() const {
+uint32_t ActionResponsePdu::getNumberOfVariableDatumRecords() const {
   return _variableDatums.size();
 }
 
@@ -90,8 +90,8 @@ void ActionResponsePdu::marshal(DataStream& dataStream) const {
   _receivingID.marshal(dataStream);
   dataStream << _requestID;
   dataStream << _requestStatus;
-  dataStream << (unsigned int)_fixedDatums.size();
-  dataStream << (unsigned int)_variableDatums.size();
+  dataStream << (uint32_t)_fixedDatums.size();
+  dataStream << (uint32_t)_variableDatums.size();
 
   for (size_t idx = 0; idx < _fixedDatums.size(); idx++) {
     FixedDatum x = _fixedDatums[idx];
@@ -162,12 +162,12 @@ int ActionResponsePdu::getMarshalledSize() const {
   marshalSize = marshalSize + 4;  // _numberOfFixedDatumRecords
   marshalSize = marshalSize + 4;  // _numberOfVariableDatumRecords
 
-  for (unsigned long long idx = 0; idx < _fixedDatums.size(); idx++) {
+  for (uint64_t idx = 0; idx < _fixedDatums.size(); idx++) {
     FixedDatum listElement = _fixedDatums[idx];
     marshalSize = marshalSize + listElement.getMarshalledSize();
   }
 
-  for (unsigned long long idx = 0; idx < _variableDatums.size(); idx++) {
+  for (uint64_t idx = 0; idx < _variableDatums.size(); idx++) {
     VariableDatum listElement = _variableDatums[idx];
     marshalSize = marshalSize + listElement.getMarshalledSize();
   }
